--- src/api/sidebar.ts
@@ -1,45 +1,0 @@
-export const siderbaritems = [
-  {
-    heading: "Administration",
-    sidebaritems: [
-      {
-        label: "Dashboard",
-        icon: "",
-      },
-
-      {
-        label: "Members",
-        icon: "",
-      },
-    ],
-  },
-  {
-    heading: "Management",
-    sidebaritems: [
-      {
-        label: "Request",
-        icon: "",
-      },
-
-      {
-        label: "Invoices",
-        icon: "",
-      },
-      {
-        label: "Transaction",
-        icon: "",
-      },
-    ],
-  },
-
-  {
-    heading: "Accounting",
-    sidebaritems: [
-      {
-        label: "Reports",
-        icon: "",
-      },
-
-    ]
-  }
-];
