+++ src/app/components/ui/sidebar/sideBar.tsx
@@ -2,31 +2,14 @@ import Image from "next/image";
 import React from "react";
 import Line from "./line/line";
 import { siderbaritems } from "@/api/sidebar";
+import Link from "next/link";
+
 
 const SideBar = () => {
   return (
-    // <div className="bg-light rounded-lg px-8 pt-[40px] pb-6 flex flex-col items-center gap-[40px]">
-    //   <div className="flex flex-col items-center">
-    //     <div className="w-[64px] h-[64px] ">
-    //       <Image
-    //         src="/profile/pp.png"
-    //         width={100}
-    //         height={100}
-    //         className="h-full w-full rounded-xl object-cover"
-    //         alt="profilePicture"
-    //       />
-    //     </div>
-    //     <div className="text-white text-md text-center  ">Your Name</div>
-    //   </div>
-    //   <Line />
-    //   <div>Profile</div>
-    //   <div>Profile</div>
-    //   <div>Profile</div>
-    // </div>
-
-    <div className="">
+    <div className=" bg-white/5 px-10 rounded-xl py-7">
       <section className="flex gap-5 flex-col items-center">
-        <div>
+        <div className="rounded-xl overflow-hidden">
           <Image
             src="/profile/pp.png"
             width={100}
@@ -38,35 +21,33 @@ const SideBar = () => {
         <p className="text-white "> Hi, Mary</p>
       </section>
 
-        <div className="py-6">
-          <Line />
-        </div>
+      <div className="py-6">
+        <Line />
+      </div>
 
       <section className="text-white">
-       
-       {
-        siderbaritems.map ((sidebarheading,index) => {
+        {siderbaritems.map((sidebarheading, index) => {
           return (
-            <div>
-            
-              {sidebarheading.heading}
+            <div className="mt-5">
+              <div> {sidebarheading.heading} </div>
 
-              {
-                sidebarheading.sidebaritems.map((sidebaritems,index) => (
-                  <div>
+              {sidebarheading.sidebaritems.map((sidebaritems, index) => (
+                <Link href={sidebaritems.link} key={index}>
+                  {" "}
+                  <div className="flex items-center gap-3 text-sm my-5 text-white/70 hover:text-white cursor-pointer ">
+                    {sidebaritems.icon}
                     {sidebaritems.label}
-                  </div>
-                )
-              )
-              }
+                  </div>{" "}
+                </Link>
+              ))}
             </div>
-          )
-        }
-      )
-       }
+          );
+          
+        })}
 
-        
       </section>
+
+      <button className="bg-cream rounded-full px-5 py-2 mt-5">Sign Out</button>
     </div>
   );
 };
